+++ NEWS.md
@@ -137,6 +137,8 @@ rowwiseDT(
 
 6. `measurev()` was implemented and documented in v1.15.0, for use within `melt()`, and it is now exported (dependent packages can now use without a NOTE from CRAN check).
 
+7. `key<-`, marked as deprecated since 2012 and unusable since v1.15.0, has been fully removed.
+
 # data.table [v1.16.2](https://github.com/Rdatatable/data.table/milestone/35) (9 October 2024)
 
 ## BUG FIXES